Rename ultralcd implementation files for consistency
